Move logger creation to another package
